+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.8.9-0.14.0] - unreleased
 ### Added
+- New command: `/commandslist` to list all client-side commands added by all installed mods
 - Chest Tracker & Analyzer:
   - added support for 'lowest BIN' prices
   - double clicking an analysis row now highlights chests that contain the clicked item